server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
rembg==2.0.50
Pillow==10.1.0 
PyJWT==2.10.1
google-generativeai==0.3.2
gunicorn
flask-swagger
flask-swagger-ui
flask-restful